Refactor, add utils.ts
