+++ src/commands/user.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode'
-import { checkLogin, login, refreshHistoryList } from './../tree/index'
+import { checkLogin, login, refreshHistoryList } from './../utils'
 import { Store } from '../store';
 import * as apiClient from '@hackmd/api';
 const API = new apiClient.default();
